[refs #85843] Refactor misc urls
